feat: validation on api based on user's session

+++ src/app/api/collections/[collectionId]/route.ts
@@ -1,5 +1,7 @@
 import { NextResponse } from "next/server";
 import prisma from "../../../../../prisma/prisma";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../../auth/[...nextauth]/route";
 
 export async function PUT(
     request: Request,
@@ -7,13 +9,20 @@ export async function PUT(
   ) {
     const { collectionId } = await params
     const body = await request.json()
+
+      const session = await getServerSession(authOptions);
+    
+      if (!session) {
+        return NextResponse.json({ error: 'You must be logged in.' }, { status: 401 })
+      }
     
     const collection = await prisma.collection.update({
       where: {
-        id: collectionId
+        id: collectionId,
+        ownerId: session.user.id
       },
       data: body
-    });
+    })
 
     if (!collection) {
       return NextResponse.json({ error: "Collection not found" }, { status: 404 })
@@ -27,11 +36,17 @@ export async function DELETE(
     { params }: { params: Promise<{ collectionId: string }> }
   ) {
     const { collectionId } = await params
+
+    const session = await getServerSession(authOptions);
     
-    // Validate if exists and if user can see it
+    if (!session) {
+      return NextResponse.json({ error: 'You must be logged in.' }, { status: 401 })
+    }
+
     const collection = await prisma.collection.delete({
       where: {
-        id: collectionId
+        id: collectionId,
+        ownerId: session.user.id
       }
     });
